pyproject.toml
[project]
name = "sam3-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "torch>=2.9.1",
    "torchvision>=0.24.0",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.7.0",
    "sam3",
    "pillow>=10.0.0",
    "numpy>=1.26,<2",
    "einops>=0.8.2",
    "triton-windows>=3.6.0.post26",
    "pycocotools>=2.0.11",
    "psutil>=7.2.2",
    "python-multipart>=0.0.24",
    "opencv-python>=4.10.0",
    "loguru>=0.7.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
]


[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }
sam3 = { path = "third_party/sam3", editable = true }

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]
docs = [
    "myst-parser>=5.0.0",
    "sphinx>=9.1.0",
    "sphinx-autodoc-typehints>=3.10.1",
    "sphinx-rtd-theme>=3.1.0",
    "sphinxcontrib-openapi>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
